requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.6
chromadb==0.4.22
pandas==2.1.4
openpyxl==3.1.2
sentence-transformers==2.3.1
tiktoken==0.5.2
